requirements.txt
pdf2image==1.16.0
Pillow==9.1.1
PyYAML==6.0
regex==2022.6.2
reportlab==3.6.10
requests==2.27.1